Update main.pdf and anexos.tex files
This commit updates the main.pdf and anexos.tex files. It includes changes to the title, authors, professor, and date in the main.pdf file. In the anexos.tex file, it updates the links to the GitHub repositories and fixes some formatting issues. The changes aim to improve the accuracy and completeness of the project documentation.

Co-authored-by: Aleu <Alesilvar@users.noreply.github.com>
Co-authored-by: KarTaGo124 <KarTaGo124@users.noreply.github.com>